requirements.txt
annotated-types==0.6.0
Authlib==1.3.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
dimod==0.12.14
diskcache==5.6.3
dwave-cloud-client==0.11.2
dwave-drivers==0.4.4
dwave-networkx==0.8.14
dwave-preprocessing==0.6.5
dwave-samplers==1.2.0
dwave-system==1.22.0
fasteners==0.19
homebase==1.0.1
idna==3.6
importlib-metadata==7.0.1
MarkupSafe==2.1.3
minorminer==0.2.13
networkx==2.8.8
numpy==1.26.3
packaging==23.2
plucky==0.4.3
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
PySocks==1.7.1
python-dateutil==2.8.2
rectangle-packer==2.0.2
requests==2.31.0
scipy==1.11.4
six==1.16.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
